requirements.txt
pandas>=2.0.0
yfinance>=0.2.30
requests>=2.31.0
python-dotenv>=1.0.0
